fix(auth): E-AUTH-REBUILD Phase1-S3 — route.ts 임의 export 제거(Next build 규약 위반)
까심 REQUEST_CHANGES(정당한 blocker): route.ts는 HTTP 메서드(GET/POST/...)+route
segment config만 export 허용 — SP_FS_COOKIE/FIREBASE_SESSION_MAX_AGE_SECONDS/
setFirebaseSessionCookie 임의 export가 next build 타입 에러(smoke-test FAIL)였다.

fix: 그 상수/헬퍼를 lib/auth/firebase-session.ts(이미 Node 런타임 전용 Firebase 세션
모듈)로 이관 — route.ts는 POST만 export하고 import해서 씀. __Host-sp_fs domain-less
회귀가드 테스트도 이관된 모듈(firebase-session.test.ts) 대상으로 유지.

검증: `next build`(Turbopack) 재실행 — route.ts 관련 에러 완전 소멸, 유일하게 남은
에러는 기존부터 있던 canvas-export.ts의 html2canvas-pro 미설치(auth와 무관, 세션
내내 확認된 pre-existing 이슈)뿐. 신규/이관 테스트 10건 전부 green.
